--- app/src/main/java/com/example/playlistmaker/EmptyListProblem.kt
@@ -1,4 +1,0 @@
-package com.example.playlistmaker
-
-class EmptyListProblem : SearchResult
-
