--- app/src/main/java/com/example/playlistmaker/EmptyListProblemViewHolder.kt
@@ -1,16 +1,0 @@
-package com.example.playlistmaker
-
-import android.view.LayoutInflater
-import android.view.ViewGroup
-import androidx.recyclerview.widget.RecyclerView
-
-class EmptyListProblemViewHolder(
-    parent: ViewGroup
-) : RecyclerView.ViewHolder(
-    LayoutInflater.from(parent.context)
-        .inflate(R.layout.activity_search_empty_list, parent, false)
-) {
-
-    fun bind(problemState: EmptyListProblem) {
-    }
-}
