--- app/src/main/java/com/example/playlistmaker/NoInternetProblem.kt
@@ -1,4 +1,0 @@
-package com.example.playlistmaker
-
-class NoInternetProblem : SearchResult
-
