+++ app/src/main/res/layout/activity_search.xml
@@ -59,45 +59,21 @@
         tools:itemCount="5"
         tools:listitem="@layout/track_view" />
 
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="1"
-        android:orientation="vertical">
+    <TextView
+        android:id="@+id/search_no_internet_problem_text"
+        style="@style/search_no_internet_problem_text"
+        tools:text="@string/no_internet_connection"
+        android:visibility="gone" />
 
-        <ImageView
-            android:id="@+id/search_no_internet_problem_image"
-            style="@style/search_no_internet_problem_image"
-            android:visibility="gone"/>
+    <Button
+        android:id="@+id/retry_search_button"
+        style="@style/search_no_internet_problem_retry_button"
+        android:visibility="gone">
+    </Button>
 
-        <TextView
-            android:id="@+id/search_no_internet_problem_text"
-            style="@style/search_problem_text"
-            tools:text="@string/no_internet_connection"
-            android:visibility="gone"/>
-
-        <Button
-            android:id="@+id/retry_search_button"
-            style="@style/search_no_internet_problem_retry_button"
-            android:visibility="gone">
-        </Button>
-    </LinearLayout>
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_weight="1"
-        android:orientation="vertical">
-
-        <ImageView
-            android:id="@+id/empty_list_problem_image"
-            style="@style/search_empty_list_problem_image"
-            android:visibility="gone"/>
-
-        <TextView
-            android:id="@+id/empty_list_problem_text"
-            style="@style/search_problem_text"
-            tools:text="@string/tracks_not_found"
-            android:visibility="gone"/>
-    </LinearLayout>
+    <TextView
+        android:id="@+id/empty_list_problem_text"
+        style="@style/empty_list_problem_text"
+        tools:text="@string/tracks_not_found"
+        android:visibility="gone" />
 </LinearLayout>
