--- app/src/main/res/layout/activity_search_empty_list.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <ImageView
-        style="@style/search_empty_list_problem_image" />
-
-    <TextView
-        style="@style/search_problem_text"
-        tools:text="@string/tracks_not_found" />
-</LinearLayout>
