--- app/src/main/res/layout/activity_search_no_internet.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <ImageView style="@style/search_no_internet_problem_image" />
-
-    <TextView
-        style="@style/search_problem_text"
-        tools:text="@string/no_internet_connection" />
-
-    <Button
-        android:id="@+id/retry_search_button"
-        style="@style/search_no_internet_problem_retry_button" />
-</LinearLayout>
